Move UvStreamTangentBitmask to new files.
